config & fixes

+++ server.py
@@ -1,7 +1,9 @@
 import urllib.request
+import configparser
 import threading
 import logging
 import socket
+import time
 import os
 
 version = 1.0
@@ -10,18 +12,38 @@ version = 1.0
 format = "%(asctime)s [%(levelname)s] %(message)s"
 logging.basicConfig(format = format, level = logging.DEBUG, datefmt = "%H:%M:%S")
 
+#config
+if os.path.isfile("server.ini"):
+    pass
+else:
+    configfile = open("server.ini", "w")
+    configfile.write("# server.ini\n" +
+                     "[Server]\n" +
+                     "port=80")
+    configfile.close()
+config = configparser.ConfigParser()
+config.read("server.ini")
+port = config["Server"]["port"]
+logging.info("config done")
+
 #htmls
 try: os.mkdir("www")
 except FileExistsError: pass
-urllib.request.urlretrieve("http://185.87.192.150/", "www/index.html")
-urllib.request.urlretrieve("http://185.87.192.150/favicon.ico", "www/favicon.ico")
+if os.path.isfile("www/index.html") and os.path.isfile("www/favicon.ico"):
+    pass
+else:
+    urllib.request.urlretrieve("http://185.87.192.150/", "www/index.html")
+    urllib.request.urlretrieve("http://185.87.192.150/favicon.ico", "www/favicon.ico")
 logging.info("www done")
 
 #error codes
 try: os.mkdir("codes") 
 except FileExistsError: pass
-urllib.request.urlretrieve("http://185.87.192.150/404_.html", "codes/404.html")
-urllib.request.urlretrieve("http://185.87.192.150/favicon.ico", "codes/favicon.ico")
+if os.path.isfile("codes/404.html") and os.path.isfile("codes/favicon.ico"):
+    pass
+else:
+    urllib.request.urlretrieve("http://185.87.192.150/404_.html", "codes/404.html")
+    urllib.request.urlretrieve("http://185.87.192.150/favicon.ico", "codes/favicon.ico")
 logging.info("codes done")
 
 #plugins
@@ -31,9 +53,9 @@ logging.info("plugins done")
 
 def start():
     server = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-    server.bind(("127.0.0.1", 80))
+    server.bind(("127.0.0.1", int(port)))
     while True:
-        server.listen(4)
+        server.listen()
         client_socket, address = server.accept()
         data = client_socket.recv(1024).decode("utf-8")
         client_socket.send(resp(data))
@@ -54,11 +76,11 @@ def resp(request_data):
                 with open("codes" + "/404.html", "rb") as file:
                     return HDRS.encode("utf-8") + file.read()
     except Exception:
-        return "Error"
+        return HDRS.encode("utf-8")
 
 Server = threading.Thread(target = start, daemon = True)
 Server.start()
-logging.info("server is running on :80")
+logging.info("server is running on :" + port)
 while True:
     cmd = input("> ")
     match cmd:
